+++ src/displaystring.ts
@@ -1,0 +1,16 @@
+export class DisplayString {
+
+  private value: string;
+  constructor(value: string) {
+
+    this.value = value;
+
+  }
+
+  toString(): string {
+
+    return this.value;
+
+  }
+
+}