+++ src/parser.ts
@@ -11,6 +11,7 @@ import {
 import { Token } from './token';
 
 import { isAscii } from './util';
+import { DisplayString } from './displaystring';
 
 export function parseDictionary(input: string): Dictionary {
 
@@ -185,6 +186,9 @@ export default class Parser {
     if (char === '@') {
       return this.parseDate();
     }
+    if (char === '%') {
+      return this.parseDisplayString();
+    }
     throw new ParseError(this.pos, 'Unexpected input');
 
   }
@@ -294,6 +298,49 @@ export default class Parser {
         outputString += char;
       }
 
+    }
+    throw new ParseError(this.pos, 'Unexpected end of input');
+
+  }
+
+  private parseDisplayString(): DisplayString {
+
+    const chars = this.getChars(2);
+    if (chars !== '%"') {
+      throw new ParseError(this.pos, 'Unexpected character. Display strings should start with %=');
+    }
+
+    const result:number[] = [];
+
+    while (!this.eof()) {
+
+      const char = this.getChar();
+      if (char.charCodeAt(0) <= 0x1F || (char.charCodeAt(0) >= 0x7F && char.charCodeAt(0) <= 0xFF)) {
+        throw new ParseError(this.pos, 'Invalid char found in DisplayString. Did you forget to escape?');
+      }
+
+      if (char==='%') {
+        const hexChars = this.getChars(2);
+        if (/^[0-9a-f]{2}$/.test(hexChars)) {
+          result.push(parseInt(hexChars, 16));
+        } else {
+          throw new ParseError(this.pos, `Unexpected sequence after % in DispalyString: "${hexChars}". Note that hexidecimals must be lowercase`);
+        }
+        continue;
+      }
+      if (char==='"') {
+        const textDecoder = new TextDecoder('utf-8', {
+          fatal: true
+        });
+        try {
+          return new DisplayString(
+            textDecoder.decode(new Uint8Array(result))
+          );
+        } catch (err) {
+          throw new ParseError(this.pos, 'Fatal error decoding UTF-8 sequence in Display String');
+        }
+      }
+      result.push(char.charCodeAt(0));
     }
     throw new ParseError(this.pos, 'Unexpected end of input');
 
@@ -428,6 +475,16 @@ export default class Parser {
   private getChar(): string {
 
     return this.input[this.pos++];
+
+  }
+  private getChars(count: number): string {
+
+    const result = this.input.substr(
+      this.pos,
+      count
+    );
+    this.pos += count;
+    return result;
 
   }
   private eof():boolean {